backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
motor==3.6.0
jinja2==3.1.4
docxtpl==0.16.7
python-docx==1.1.2
starlette==0.38.2
itsdangerous==2.2.0
email-validator==2.2.0
# Pure Python PDF generation (no Word/LibreOffice needed)
reportlab==4.2.5
python-docx-template==0.1.3
